registration form changes #day2 Added signals changes in forms
