+++ users/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import Doctor, Patient, Visit, Documentation, Medicine, Disease, Address
+from main.models import Doctor, Patient, Visit, Documentation, Medicine, Disease, Address
 # Register your models here.
 
 admin.site.register(Doctor)
